Test Code Improvements
In `Experimental.System.Messaging.Test.csproj`, we upgraded .NET Core 2.0 to .NET 8.0 and updated the test-related package versions to the latest.

In `QueueClientTest.cs`, we modified the function to check for the existence of a queue and create one if necessary, ensuring test independence. We also separated the queue path into a variable for improved readability.